VMR vs MSFT signing comparison
