+++ src/sdk/eng/GatherDrops.ps1
@@ -9,14 +9,17 @@ param(
   [Parameter(Mandatory=$true)]
   [String]$githubPat,
   [Parameter(Mandatory=$true)]
-  [String]$azdevPat
+  [String]$azdevPat,
+  [Parameter(Mandatory=$false)]
+  [String]$assetFilter = ".*",
+  [Switch]$nonShipping = $false
 )
 $jsonContent = Get-Content -Path $filePath -Raw | ConvertFrom-Json
 foreach ($repo in $jsonContent.repositories) {
     $remoteUri = $repo.remoteUri
     $commitSha = $repo.commitSha
     $path = "$outputPath$($repo.path)"
-    $darcCommand = "$darcPath gather-drop -c $commitSha -r $remoteUri --non-shipping --skip-existing --continue-on-error --use-azure-credential-for-blobs -o $path --github-pat $githubPat --azdev-pat $azdevPat --verbose --ci"
+    $darcCommand = "$darcPath gather-drop -c $commitSha -r $remoteUri --skip-existing --continue-on-error --use-azure-credential-for-blobs -o $path --github-pat $githubPat --azdev-pat $azdevPat --asset-filter $assetFilter --verbose --ci --include-released $(if ($nonShipping) { '--non-shipping' })"
     Write-Output "Gathering drop for $remoteUri"
     Invoke-Expression $darcCommand
 }
