+++ src/sdk/eng/vmr-msft-comparison-baseline.json
@@ -511,5 +511,17 @@
     "idMatch": "VS\\.Redist\\..*",
     "descriptionMatch": "ProjectUrl",
     "justification": "VS redist packages do not ship and should have their project url set to the VMR to identify where they came from."
+  },
+  {
+    "issueType": "Unsigned",
+    "idMatch": ".*/dotnet-sdk-pgo-.*.zip",
+    "descriptionMatch": ".* is Excluded in the VMR but should be Signed",
+    "justification": "The VMR build purposely excludes signing the PGO builds."
+  },
+  {
+    "issueType": "Unsigned",
+    "idMatch": ".*/dotnet-sdk-pgo-.*.tar.gz",
+    "descriptionMatch": ".* is Excluded in the VMR but should be Signed",
+    "justification": "The VMR build purposely excludes signing the PGO builds."
   }
 ]
